+++ R/show_models.R
@@ -25,7 +25,7 @@
 show_models <- function(models, model_names = NULL, covariates = NULL,
                         merge_models = FALSE, drop_controls = FALSE, ...) {
   stopifnot(inherits(models, "ezcox_models") | all(sapply(models, function(x) inherits(x, "coxph"))))
-  pkg_version = packageVersion("forestmodel")
+  pkg_version <- packageVersion("forestmodel")
   if (pkg_version$major == 0 & pkg_version$minor < 6) {
     message("Please install the recent version of forestmodel firstly.")
     message("Run the following command:")