+++ NEWS.md
@@ -1,3 +1,8 @@
+# ezcox 0.4.0
+
+* Add function `show_forest`, which is a wrapper of
+function `ezcox`, `get_models` and `show_models`.
+
 # ezcox 0.3.1
 
 * Fix the bug that a `NA` is added to model names when no control